[AIC-py][eval] Move openai model graded eval tests to its own file
Just organizing
